fix(onboard): separate portable gateway and registry (#9461)
