+++ docs/inference/set-up-openai-compatible-endpoint.mdx
@@ -132,6 +132,11 @@ curl -fsSL https://www.nvidia.com/nemoclaw.sh | NEMOCLAW_AGENT=langchain-deepage
 
 </AgentOnly>
 
+The portable profile creates the `openshell-docker` network on `169.254.1.0/24`.
+On that network, `169.254.1.2` is the host-gateway address for `host.openshell.internal` and carries authenticated mTLS plus sandbox-JWT callbacks to OpenShell on port `8080`.
+The managed local registry uses the distinct address `169.254.1.3` on port `5000`.
+Keeping these addresses distinct prevents registry traffic from intercepting gateway callbacks.
+
 The portable profile handles the descriptor as follows:
 
 | Descriptor state | Onboarding result |
