+++ src/lib/onboard/experimental/portable-profile.ts
@@ -6,6 +6,7 @@ import { DEFAULT_DOCKER_DRIVER_NETWORK_NAME } from "./docker-network-authority";
 export const EXPERIMENTAL_PROFILE_ENV = "NEMOCLAW_EXPERIMENTAL_PROFILE";
 export const PORTABLE_EXPERIMENTAL_PROFILE = "portable";
 export const PORTABLE_HOST_GATEWAY_IP = "169.254.1.2";
+export const PORTABLE_REGISTRY_IP = "169.254.1.3";
 export const PORTABLE_DOCKER_NETWORK_NAME = DEFAULT_DOCKER_DRIVER_NETWORK_NAME;
 export const PORTABLE_DOCKER_NETWORK_SUBNET = "169.254.1.0/24";
 export const PORTABLE_LOCAL_REGISTRY = "localhost:5000";
